requirements.txt
pyinstaller>=6.18.0
pyserial>=3.5
pillow>=12.1.1